Issue #71: Seperate subroker apps
