+++ backend/backend/urls.py
@@ -17,6 +17,10 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('api/', include('subroker.urls')),
+    path('api/user/', include('user.urls')),
+    path('api/group/', include('group.urls')),
+    path('api/content/', include('content.urls')),
+    path('api/ott/', include('ott.urls')),
+    path('api/review/', include('review.urls')),
     path('admin/', admin.site.urls),
 ]
