+++ backend/content/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from content.models import Content
+
+# Register your models here.
+admin.site.register(Content)
