+++ backend/content/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class SubrokerConfig(AppConfig):
+class ContentConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'subroker'
+    name = 'content'
