+++ backend/group/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from group.models import Group
+
+# Register your models here.
+admin.site.register(Group)
