+++ backend/ott/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from ott.models import Ott
+
+# Register your models here.
+admin.site.register(Ott)
