+++ backend/review/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from review.models import Review
+
+# Register your models here.
+admin.site.register(Review)
