+++ backend/review/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    path('<int:review_id>/', views.review_detail)
+]
